backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
josepy==1.14.0
mozilla-django-oidc==4.0.1
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
schema==0.7.7
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
